+++ integration_tests/sdk/aqueduct_tests/delete_workflow_test.py
@@ -1,18 +1,12 @@
 import pandas as pd
 import pytest
 from aqueduct.constants.enums import LoadUpdateMode
-from aqueduct.error import InvalidRequestError
-from data_objects import DataObject
-from relational import SHORT_SENTIMENT_SQL_QUERY, all_relational_DBs
-from utils import (
-    check_flow_doesnt_exist,
-    check_table_doesnt_exist,
-    check_table_exists,
-    extract,
-    generate_table_name,
-    publish_flow_test,
-    save,
-)
+from aqueduct.error import InvalidRequestError, InvalidUserArgumentException
+
+from ..shared.data_objects import DataObject
+from ..shared.relational import SHORT_SENTIMENT_SQL_QUERY, all_relational_DBs
+from ..shared.utils import extract, generate_table_name, publish_flow_test
+from .save import save
 
 
 def test_delete_workflow_invalid_saved_objects(client, flow_name, data_integration, engine):
@@ -46,32 +40,32 @@ def test_force_delete_workflow_saved_objects(
 ):
     """Check the flow with object(s) saved with update_mode=APPEND can only be deleted if in force mode."""
     table_name = generate_table_name()
-    table = data_integration.sql(query=SHORT_SENTIMENT_SQL_QUERY)
-    save(data_integration, table, name=table_name, update_mode=LoadUpdateMode.REPLACE)
+    table_artifact = data_integration.sql(query=SHORT_SENTIMENT_SQL_QUERY)
+    save(data_integration, table_artifact, name=table_name, update_mode=LoadUpdateMode.REPLACE)
 
     flow = publish_flow_test(
         client,
-        table,
+        table_artifact,
         name=flow_name(),
         engine=engine,
     )
 
-    save(data_integration, table, name=table_name, update_mode=LoadUpdateMode.APPEND)
+    save(data_integration, table_artifact, name=table_name, update_mode=LoadUpdateMode.APPEND)
     flow = publish_flow_test(
         client,
-        table,
+        table_artifact,
         engine=engine,
         existing_flow=flow,
     )
 
-    extracted_table_data = table.get()
-    validator.check_saved_artifact(
+    extracted_table_data = table_artifact.get()
+    validator.check_saved_artifact_data(
         flow,
-        table.id(),
+        table_artifact.id(),
         expected_data=pd.concat([extracted_table_data, extracted_table_data], ignore_index=True),
     )
 
-    tables = client.flow(flow.id()).list_saved_objects()
+    tables = flow.list_saved_objects()
     assert table_name in [item.spec.identifier() for item in tables[data_integration]]
 
     # Doesn't work if don't force because it is created in append mode.
@@ -80,10 +74,13 @@ def test_force_delete_workflow_saved_objects(
 
     # Actually delete the flow.
     client.delete_flow(flow.id(), saved_objects_to_delete=tables, force=True)
-    check_flow_doesnt_exist(client, flow.id())
+
+    # Check that the flow can not longer be accessed.
+    with pytest.raises(InvalidUserArgumentException, match="Unable to find a flow"):
+        client.flow(flow.id())
 
     # Check table no longer exists
-    check_table_doesnt_exist(data_integration, table_name)
+    validator.check_saved_artifact_data_does_not_exist(table_artifact.id())
 
 
 @pytest.mark.enable_only_for_data_integration_type(*all_relational_DBs())
@@ -97,36 +94,32 @@ def test_delete_workflow_saved_objects_twice(
     """
     table_name = generate_table_name()
 
-    table = data_integration.sql(query=SHORT_SENTIMENT_SQL_QUERY)
-    save(data_integration, table, name=table_name, update_mode=LoadUpdateMode.REPLACE)
+    table_artifact = data_integration.sql(query=SHORT_SENTIMENT_SQL_QUERY)
+    save(data_integration, table_artifact, name=table_name, update_mode=LoadUpdateMode.REPLACE)
 
     # Workflow 1's name not specified, so given a random workflow name.
     flow1 = publish_flow_test(
         client,
-        table,
+        table_artifact,
         name=flow_name(),
         engine=engine,
     )
 
     # Workflow 2's name not specified, so given a random workflow name.
-    save(data_integration, table, name=table_name, update_mode=LoadUpdateMode.APPEND)
+    save(data_integration, table_artifact, name=table_name, update_mode=LoadUpdateMode.APPEND)
     flow2 = publish_flow_test(
         client,
-        table,
+        table_artifact,
         name=flow_name(),
         engine=engine,
     )
 
-    extracted_table_data = table.get()
-    validator.check_saved_artifact(
+    extracted_table_data = table_artifact.get()
+    validator.check_saved_artifact_data(
         flow1,
-        table.id(),
+        table_artifact.id(),
         expected_data=pd.concat([extracted_table_data, extracted_table_data], ignore_index=True),
     )
-
-    # Check table is properly created at the integration.
-    # Need to poll initially in case still writing table.
-    check_table_exists(data_integration, table_name)
 
     tables = client.flow(flow1.id()).list_saved_objects()
     tables_1 = set([item.spec.identifier() for item in tables[data_integration]])
@@ -140,8 +133,10 @@ def test_delete_workflow_saved_objects_twice(
     client.delete_flow(flow1.id(), saved_objects_to_delete=tables, force=True)
 
     # Check flow indeed deleted and that the table no longer exists.
-    check_flow_doesnt_exist(client, flow1.id())
-    check_table_doesnt_exist(data_integration, table_name)
+    with pytest.raises(InvalidUserArgumentException, match="Unable to find a flow"):
+        client.flow(flow1.id())
+
+    validator.check_saved_artifact_data_does_not_exist(table_artifact.id())
 
     # Try to delete table deleted by other flow.
     with pytest.raises(Exception) as e_info:
@@ -149,4 +144,5 @@ def test_delete_workflow_saved_objects_twice(
     assert str(e_info.value).startswith("Failed to delete")
 
     # Failed to delete tables, but flow should be removed.
-    check_flow_doesnt_exist(client, flow2.id())
+    with pytest.raises(InvalidUserArgumentException, match="Unable to find a flow"):
+        client.flow(flow2.id())
