+++ integration_tests/sdk/aqueduct_tests/error_handling_test.py
@@ -1,11 +1,11 @@
 import pytest
 from aqueduct.error import AqueductError, InvalidUserArgumentException
-from data_objects import DataObject
-from relational import all_relational_DBs
-from utils import extract
 
 import aqueduct
 from aqueduct import op
+
+from ..shared.data_objects import DataObject
+from ..shared.utils import extract
 
 
 @aqueduct.op()
@@ -14,20 +14,8 @@ def bad_op(df):
     return df
 
 
-GOOD_QUERY = "SELECT * FROM hotel_reviews"
-BAD_QUERY = "SELEC * FROM sdafawefa"
-
 # These tips should match executor code so that we can verify the correct error is generated.
-TIP_EXTRACT = "We couldn't execute the provided query. Please double check your query is correct."
 TIP_OP_EXECUTION = "Error executing operator. Please refer to the stack trace for fix."
-
-
-@pytest.mark.enable_only_for_data_integration_type(*all_relational_DBs())
-def test_handle_relational_query_error(client, data_integration):
-    try:
-        _ = data_integration.sql(query=BAD_QUERY)
-    except AqueductError as e:
-        assert TIP_EXTRACT in e.message
 
 
 def test_handle_bad_op_error(client, data_integration):
