+++ integration_tests/sdk/aqueduct_tests/local_test.py
@@ -1,10 +1,11 @@
-from checks_test import success_on_single_table_input
-from data_objects import DataObject
 from pandas import DataFrame
 from pandas._testing import assert_frame_equal
-from test_functions.simple.model import dummy_sentiment_model, dummy_sentiment_model_multiple_input
-from test_metrics.constant.model import constant_metric
-from utils import extract
+
+from ..shared.data_objects import DataObject
+from ..shared.utils import extract
+from .checks_test import success_on_single_table_input
+from .test_functions.simple.model import dummy_sentiment_model, dummy_sentiment_model_multiple_input
+from .test_metrics.constant.model import constant_metric
 
 
 def test_local_operator(client, data_integration):
