+++ integration_tests/sdk/aqueduct_tests/table_artifact_test.py
@@ -2,10 +2,11 @@ import math
 import time
 
 import pandas as pd
-from data_objects import DataObject
-from utils import extract
 
 from aqueduct import op
+
+from ..shared.data_objects import DataObject
+from ..shared.utils import extract
 
 
 def test_great_expectations_check(client, data_integration):
