+++ integration_tests/sdk/aqueduct_tests/test_function.py
@@ -1,3 +1,4 @@
+# TODO(ENG-1470): Currently unused until test is fixed.
 def dummy_sentiment_model_function(df):
     df["positivity"] = 123
     return df
