+++ integration_tests/sdk/aqueduct_tests/test_metrics/constant/model.py
