+++ integration_tests/sdk/shared/demo_db.py
@@ -1,0 +1,15 @@
+from typing import List
+
+
+def demo_db_tables() -> List[str]:
+    return [
+        "customer_activity",
+        "customers",
+        "hotel_reviews",
+        "mpg",
+        "tips",
+        "wine",
+        "diabetes",
+        "house_prices",
+        "pred_churn",
+    ]
